Move definition of eigendecomposition to MooseTypes, simplify input files
Clean up existing hierarchy
