+++ modules/tensor_mechanics/include/materials/ADGeneralizedRadialReturnStressUpdate.h
@@ -13,26 +13,6 @@
 #include "ADGeneralizedReturnMappingSolution.h"
 #include "MooseTypes.h"
 
-#include <Eigen/Core>
-#include <Eigen/Eigenvalues>
-#include <Eigen/Dense>
-#include "metaphysicl/dualnumber.h"
-
-namespace Eigen
-{
-namespace internal
-{
-template <>
-struct cast_impl<ADReal, int>
-{
-  static inline int run(const ADReal & x) { return static_cast<int>(MetaPhysicL::raw_value(x)); }
-};
-} // namespace internal
-} // namespace Eigen
-
-typedef Eigen::Matrix<ADReal, 6, 6, Eigen::DontAlign> AnisotropyMatrix;
-typedef Eigen::Matrix<Real, 6, 6, Eigen::DontAlign> AnisotropyMatrixReal;
-
 /**
  * ADGeneralizedRadialReturnStressUpdate computes the generalized radial return stress increment for
  * anisotropic (Hill-like) creep and plasticity. This generalized radial return mapping class
@@ -114,10 +94,9 @@ protected:
    * @param effective_trial_stress Effective trial stress
    * @param elasticityTensor     Elasticity tensor
    */
-  virtual void computeStressInitialize(const ADDenseVector & /*effective_trial_stress*/,
-                                       const ADRankFourTensor & /*elasticity_tensor*/)
-  {
-  }
+  virtual void computeStressInitialize(const ADDenseVector & /*stress_dev*/,
+                                       const ADDenseVector & /*stress*/,
+                                       const ADRankFourTensor & /*elasticity_tensor*/) = 0;
 
   /**
    * Perform any necessary steps to finalize state after return mapping iterations
@@ -144,9 +123,6 @@ protected:
 
   void outputIterationSummary(std::stringstream * iter_output,
                               const unsigned int total_it) override;
-
-  /// 2 * shear modulus
-  ADReal _two_shear_modulus;
 
   /// Equivalent creep/plastic strain
   ADMaterialProperty<Real> & _effective_inelastic_strain;
