+++ modules/tensor_mechanics/include/materials/ADTransverselyIsotropicCreepStressUpdate.h
@@ -39,7 +39,8 @@ public:
                                  const ADMaterialProperty<RankTwoTensor> & strain_rate) override;
 
 protected:
-  virtual void computeStressInitialize(const ADDenseVector & effective_trial_stress,
+  virtual void computeStressInitialize(const ADDenseVector & stress_dev,
+                                       const ADDenseVector & stress,
                                        const ADRankFourTensor & elasticity_tensor) override;
   virtual ADReal computeResidual(const ADDenseVector & effective_trial_stress,
                                  const ADDenseVector & stress_new,
@@ -112,4 +113,7 @@ protected:
 
   /// Square of the q function for orthotropy
   ADReal _qsigma;
+
+  /// 2 * shear modulus
+  ADReal _two_shear_modulus;
 };
