+++ modules/tensor_mechanics/test/tests/ad_anisotropic_plasticity/ad_aniso_plasticity_y.i
@@ -33,8 +33,6 @@
 [GlobalParams]
   displacements = 'disp_x disp_y disp_z'
   volumetric_locking_correction = true
-  order = FIRST
-  family = LAGRANGE
 []
 
 [AuxVariables]
@@ -42,15 +40,15 @@
     order = CONSTANT
     family = MONOMIAL
   []
-  [plasticity_strain_xx]
+  [plastic_strain_xx]
     order = CONSTANT
     family = MONOMIAL
   []
-  [plasticity_strain_xy]
+  [plastic_strain_xy]
     order = CONSTANT
     family = MONOMIAL
   []
-  [plasticity_strain_yy]
+  [plastic_strain_yy]
     order = CONSTANT
     family = MONOMIAL
   []
@@ -63,14 +61,7 @@
     family = MONOMIAL
   []
 []
-[Variables]
-  [disp_x]
-  []
-  [disp_y]
-  []
-  [disp_z]
-  []
-[]
+
 [AuxKernels]
   [hydrostatic_stress]
     type = ADRankTwoScalarAux
@@ -80,22 +71,22 @@
   []
   [plasticity_strain_xx]
     type = ADRankTwoAux
-    rank_two_tensor = trial_plasticity_plasticity_strain
-    variable = plasticity_strain_xx
+    rank_two_tensor = trial_plasticity_plastic_strain
+    variable = plastic_strain_xx
     index_i = 0
     index_j = 0
   []
   [plasticity_strain_xy]
     type = ADRankTwoAux
-    rank_two_tensor = trial_plasticity_plasticity_strain
-    variable = plasticity_strain_xy
+    rank_two_tensor = trial_plasticity_plastic_strain
+    variable = plastic_strain_xy
     index_i = 0
     index_j = 1
   []
   [plasticity_strain_yy]
     type = ADRankTwoAux
-    rank_two_tensor = trial_plasticity_plasticity_strain
-    variable = plasticity_strain_yy
+    rank_two_tensor = trial_plasticity_plastic_strain
+    variable = plastic_strain_yy
     index_i = 1
     index_j = 1
   []
@@ -126,9 +117,8 @@
 [Modules/TensorMechanics/Master]
   [all]
     strain = FINITE
-    generate_output = 'elastic_strain_xx elastic_strain_yy elastic_strain_xy stress_xx stress_xy '
-                      'stress_yy'
     use_automatic_differentiation = true
+    add_variables = true
   []
 []
 
@@ -247,7 +237,7 @@
   []
   [plasticity_strain_yy]
     type = ElementalVariableValue
-    variable = plasticity_strain_yy
+    variable = plastic_strain_yy
     execute_on = 'TIMESTEP_END'
     elementid = 39
   []
